Add confirmation dialog box for delete function

+++ components/org.wso2.carbon.siddhi.editor.core/src/main/resources/web/js/design-view/design-grid.js
@@ -919,17 +919,59 @@ define(['require', 'log', 'jquery', 'backbone', 'lodash', 'designViewUtils', 'dr
                     var close_icon_overlay = connectionObject.addOverlay([
                         "Custom", {
                             create: function () {
-                                return $('<span><i class="fw fw-delete"></i></span>');
+                                return $(
+                                    '<span><i class="fw fw-delete" id="' + connectionObject.id +
+                                    '"data-toggle="popover"></i></span>');
                             },
                             location: 0.60,
                             id: "close",
                             events: {
                                 click: function () {
-                                    self.jsPlumbInstance.deleteConnection(connectionObject);
+                                    popOverForConnector();
                                 }
                             }
                         }
                     ]);
+
+                    function popOverForConnector() {
+                        $('#' + connectionObject.id).popover({
+                            trigger: 'focus',
+                            title: 'Are you sure you want to delete?',
+                            html: true,
+                            content: function () {
+                                return $('.pop-over').html();
+                            }
+                        });
+                        $('#' + connectionObject.id).popover("show");
+                        $(".overlayed-container ").fadeTo(200, 1);
+                        // Custom jQuery to hide popover on click of the close button
+                        $("#" + connectionObject.id).siblings(".popover").on("click", ".popover-footer .btn.yes",
+                            function () {
+                                if (connectionObject.connector !== null) {
+                                    self.jsPlumbInstance.deleteConnection(connectionObject);
+                                }
+                                $(".overlayed-container ").fadeOut(200);
+                                $(this).parents(".popover").popover('hide');
+                            });
+                        $("#" + connectionObject.id).siblings(".popover").on("click", ".popover-footer .btn.no",
+                            function () {
+                                $(".overlayed-container ").fadeOut(200);
+                                $(this).parents(".popover").popover('hide');
+                                close_icon_overlay.setVisible(false);
+                            });
+                        // Dismiss the pop-over by clicking outside
+                        $('.overlayed-container ').off('click');
+                        $('.overlayed-container ').on('click', function (e) {
+                            $('[data-toggle="popover"]').each(function () {
+                                if (!$(this).is(e.target) && $(this).has(e.target).length === 0 && $('.popover').has(
+                                    e.target).length === 0) {
+                                    $(this).popover('hide');
+                                    $(".overlayed-container ").fadeOut(200);
+                                    close_icon_overlay.setVisible(false);
+                                }
+                            });
+                        });
+                    }
 
                     if (isConnectionMadeInsideAPartition) {
                         close_icon_overlay.setVisible(false);
@@ -939,7 +981,9 @@ define(['require', 'log', 'jquery', 'backbone', 'lodash', 'designViewUtils', 'dr
                         });
                         // hide the close icon when the mouse is not on the connection path
                         connectionObject.bind('mouseleave', function () {
-                            close_icon_overlay.setVisible(false);
+                            if ($("#" + connectionObject.id).siblings(".popover").length == 0) {
+                                close_icon_overlay.setVisible(false);
+                            }
                         });
                     } else {
                         close_icon_overlay.setVisible(false);
@@ -949,7 +993,9 @@ define(['require', 'log', 'jquery', 'backbone', 'lodash', 'designViewUtils', 'dr
                         });
                         // hide the close icon when the mouse is not on the connection path
                         connectionObject.bind('mouseout', function () {
-                            close_icon_overlay.setVisible(false);
+                            if ($("#" + connectionObject.id).siblings(".popover").length == 0) {
+                                close_icon_overlay.setVisible(false);
+                            }
                         });
                     }
                 });
